Move base logger
